requirements.txt
python-telegram-bot==20.6
openai
deep-translator
Pillow
pytube==15.0.0
flask==2.3.3
python-dotenv==1.0.0
serpapi